Add files via upload
This 3D project was built by a group of two people, and we use threejs library for this project. The initial concept was using the cell as a unit to explain a circle of life for human. User can scroll the mouse to experience this game, and we also chose a background music to make user immerse this scenario.
Enjoy!